requirements.txt
"""Requirements."""
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
litellm>=1.0.0
httpx>=0.25.0
parsel>=1.8.0
loguru>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
redis>=5.0.0
aiosqlite>=0.19.0
playwright>=1.40.0
